+++ src/routes/api/performance/[id]/+server.ts
@@ -1,6 +1,7 @@
 import { type PerformanceInterface, selectInstrument} from "$lib/server/common";
 import {deleteById, queryTable, updatePerformance} from "$lib/server/db";
 import {json} from "@sveltejs/kit";
+import { isAuthorized } from '$lib/server/apiAuth';
 
 export async function GET({params, request}) {
     try {
@@ -14,6 +15,10 @@ export async function GET({params, request}) {
     }
 }
 export async function PUT({params, request}) {
+    // Get the Authorization header
+    if (!isAuthorized(request.headers.get('Authorization'))) {
+        return new Response('Unauthorized', { status: 401 });
+    }
     try {
         // the following fields are often not included
         // order, concert_time, warm_up_room_name, warm_up_room_start, warm_up_room_name
@@ -75,6 +80,10 @@ export async function PUT({params, request}) {
 }
 
 export async function DELETE({params, request}){
+    // Get the Authorization header
+    if (!isAuthorized(request.headers.get('Authorization'))) {
+        return new Response('Unauthorized', { status: 401 });
+    }
     const rowCount = await deleteById('performance', params.id);
 
     if (rowCount != null && rowCount > 0) {
